Revamp the display of attachments.
Attachments are now shown in a table.

I removed the "insert content" link, since that generates text for
which there is no formatter. I also removed the associated
"insert_content" action in Controller::Attachment.

I also made the "view in browser" link only show up for text and
images. For images, it also shows a thumbnail inline in the table.

I removed the "inline" link entirely, since it only worked for images,
and just showed an 800x600 image. Modern browsers will scale a
too-big-image to fit in the window anyway, so this is pointless. I
left the action, though, because it may be useful elsewhere.