Base the edit preview & textarea height on the window size.
This change means that the preview & textarea should fit in the
viewport (maybe with a little scrolling initially).

This also changes the vertical split mode so that the preview and
textarea are nearly the same height which looks nicer, and adds a
small margin separating the two in vertical split mode.