Name test databases with process ID for parallelism
This change allows the t/t/schema_DBIC*.t tests to be run in parallel by
using the process ID in the temporary database file name.

The original tests have some code for tidying up the temporary files but
it's all commented out, so I just added the appropriate .gitignore entry
to ignore the temporary files and left it at that.